requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
cffi==1.15.1
cryptography==41.0.1
Django==4.2.2
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pycparser==2.21
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
redis==4.5.5
sqlparse==0.4.4
typing_extensions==4.6.3
